bidir map works, need to test component removal

+++ Grok3d_Windows/inc/grok3d_types.h
@@ -62,7 +62,7 @@ namespace Grok3d
     {
         class GRK_EntityHandle;
 
-        typedef int GRK_Entity;
+        typedef size_t GRK_Entity;
     }
 
     namespace Components
