+++ Grok3d_Windows/src/Entity/EntityHandle.cpp
@@ -49,7 +49,7 @@ bool GRK_EntityHandle::operator==(const Grok3d::Entities::GRK_EntityHandle& rhs)
 
 typename std::hash<GRK_EntityHandle>::result_type std::hash<GRK_EntityHandle>::operator()(std::hash<GRK_EntityHandle>::argument_type const& e) const
 {
-    return hash<int>{}(e.m_entity);
+    return hash<size_t>{}(e.m_entity);
 }
 
 bool Grok3d::Entities::operator==(const int entity, const GRK_EntityHandle& handle)
